Move test to separate Quarkus app to avoid confusion
