+++ google-cloud-graalvm-samples/graalvm-samples-quarkus/quarkus-integration-test/src/test/java/com/example/NativeExampleResourceIT.java
@@ -23,10 +23,19 @@ import io.quarkus.test.junit.NativeImageTest;
 import java.io.IOException;
 import org.junit.jupiter.api.Test;
 
-/**
- * This executes the same tests in {@link QuarkusPubsubSampleApplicationTest} but in native image
- * mode.
- */
 @NativeImageTest
-public class NativeQuarkusPubsubSampleApplicationIT extends QuarkusPubsubSampleApplicationTest {
+public class NativeExampleResourceIT extends ExampleResourceTest {
+
+  @Test
+  public void validateHeaders() throws IOException {
+    String headers =
+        given()
+            .when().get("/headers")
+            .body().print();
+
+    String[] headerTokens = headers.split(" ");
+    assertThat(headerTokens).hasSizeGreaterThan(0);
+    assertThat(headerTokens[0]).startsWith("gl-java");
+    assertThat(headerTokens[0]).endsWith("-graalvm");
+  }
 }
