+++ google-cloud-graalvm-support/src/main/resources/META-INF/native-image/com.google.cloud/google-cloud-graalvm-support/native-image.properties
@@ -8,6 +8,4 @@ Args = -H:+AllowIncompleteClasspath --report-unsupported-elements-at-runtime \
     io.grpc.netty.shaded.io.netty.internal.tcnative.SSLPrivateKeyMethod,\
     io.grpc.netty.shaded.io.grpc.netty,\
     io.grpc.netty.shaded.io.netty.channel.epoll,\
-    io.grpc.netty.shaded.io.netty.channel.unix,\
-    com.google.api.gax.core.GaxProperties,\
-    com.google.api.gax.grpc.GaxGrpcProperties
+    io.grpc.netty.shaded.io.netty.channel.unix
